requirements.txt
django==3.2.6
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-htmlmin==0.11.0
django-split-settings==1.0.1
django-types==0.7.0
gunicorn==20.1.0
mysqlclient==2.0.3
passlib==1.7.4
python-dateutil==2.8.2
envtoml==0.1.2
# toml==0.10.2              # Installed as dependency for envtoml
# asgiref==3.4.1            # Installed as dependency for django
# beautifulsoup4==4.9.3     # Installed as dependency for django-htmlmin
# html5lib==1.1             # Installed as dependency for django-htmlmin
# pytz==2021.1              # Installed as dependency for django
# six==1.16.0               # Installed as dependency for html5lib
# soupsieve==2.2.1          # Installed as dependency for beautifulsoup4
# sqlparse==0.4.1           # Installed as dependency for django, django-debug-toolbar
# webencodings==0.5.1       # Installed as dependency for html5lib
